Module single_thread_server::server::http::body

body module is responsible for how HTTP body are parsed and stored

Re-exports

pub use value::Application;

pub use value::Audio;

pub use value::Image;

pub use value::Multipart;

pub use value::Text;

pub use value::Video;

Modules

value

value module defines the specific Enums for each of the ContentType variants

Structs

Body

Body struct defines the structure of an HTTP body

Enums

ContentType

ContentType enum define the different types types of content types

Traits

VariantName

VariantName trait defines method to return string of variant names