Module single_thread_server::server::http::request::request

request module is responsible for functions parsing HTTP requests

Functions

get_data

get_data method extras metadata and possible Body from request

get_key_value_pair

get_key_value_pair extracts key and associated value from unparsed metadata string

get_start_line

get_start_line method extracts the method, http version and target from the first line of a request

parse

parse method converts raw http bytes into Method enum